+++ src/renderer/menu/menu.ts
@@ -19,8 +19,23 @@ import { InputState, InputTracker } from './input-tracker';
 import { MenuTheme } from './menu-theme';
 import { closestEquivalentAngle } from '../math';
 
-const CENTER_RADIUS = 50;
-const PARENT_DISTANCE = 150;
+/** These options can be given to the constructor of the menu. */
+export class MenuOptions {
+  /** Clicking inside this radius will select the parent element. */
+  centerDeadZone = 50;
+
+  /**
+   * The distance in pixels at which the parent menu item is placed if a submenu is
+   * selected close to the parent.
+   */
+  minParentDistance = 150;
+
+  /** The time in milliseconds it takes to fade in the menu. */
+  fadeInDuration = 150;
+
+  /** The time in milliseconds it takes to fade out the menu. */
+  fadeOutDuration = 200;
+}
 
 /**
  * The menu is the main class of Kando. It stores a tree of items which is used to render
@@ -54,10 +69,16 @@ export class Menu extends EventEmitter {
   private root: IRenderedMenuItem = null;
 
   /**
+   * This holds some global options for the menu. These options can be set when the menu
+   * is created and will be used to configure the menu's behavior.
+   */
+  private options: MenuOptions = null;
+
+  /**
    * This holds some information which is passed to the menu when it is shown from the
    * main process. For instance, it holds the window size and the initial mouse position.
    */
-  private options: IShowMenuOptions;
+  private showMenuOptions: IShowMenuOptions;
 
   /**
    * The hovered item is the menu item which is currently hovered by the mouse. It is used
@@ -95,6 +116,9 @@ export class Menu extends EventEmitter {
    */
   private input: InputTracker = new InputTracker();
 
+  /** This timeout is used to clear the menu div after the fade-out animation. */
+  private hideTimeout: NodeJS.Timeout;
+
   /**
    * The constructor will attach event listeners to the given container element. It will
    * also initialize the input tracker and the gesture detection.
@@ -104,11 +128,28 @@ export class Menu extends EventEmitter {
    */
   constructor(
     private container: HTMLElement,
-    private theme: MenuTheme
+    private theme: MenuTheme,
+    options: Partial<MenuOptions> = {}
   ) {
     super();
 
     this.container = container;
+    this.options = { ...new MenuOptions(), ...options };
+
+    // Store the fade-in and fade-out durations as CSS variables.
+    CSS.registerProperty({
+      name: '--fade-in-duration',
+      syntax: '<time>',
+      inherits: false,
+      initialValue: `${this.options.fadeInDuration}ms`,
+    });
+
+    CSS.registerProperty({
+      name: '--fade-out-duration',
+      syntax: '<time>',
+      inherits: false,
+      initialValue: `${this.options.fadeOutDuration}ms`,
+    });
 
     // When the mouse is moved, we store the absolute mouse position, as well as the mouse
     // position, distance, and angle relative to the currently selected item.
@@ -151,7 +192,7 @@ export class Menu extends EventEmitter {
       if (
         this.input.state === InputState.eClicked &&
         this.selectionChain.length === 1 &&
-        this.input.distance < CENTER_RADIUS
+        this.input.distance < this.options.centerDeadZone
       ) {
         this.emit('cancel');
         return;
@@ -250,7 +291,7 @@ export class Menu extends EventEmitter {
       // children in marking mode in order to prevent unwanted actions. This way the user
       // can always check if the correct action was selected before executing it.
       if (
-        !this.options.anchoredMode &&
+        !this.showMenuOptions.anchoredMode &&
         this.draggedItem &&
         this.draggedItem.children?.length > 0
       ) {
@@ -269,16 +310,23 @@ export class Menu extends EventEmitter {
    * given root item and all its children. It will also set up the angles and positions of
    * all items and show the menu.
    *
-   * @param options Some additional information on how to show the menu.
+   * @param showMenuOptions Some additional information on how to show the menu.
    */
-  public show(root: IRenderedMenuItem, options: IShowMenuOptions) {
+  public show(root: IRenderedMenuItem, showMenuOptions: IShowMenuOptions) {
+    // Cancel any ongoing hiding.
+    if (this.hideTimeout) {
+      clearTimeout(this.hideTimeout);
+      this.hideTimeout = null;
+    }
+
     this.clear();
 
-    this.options = options;
+    this.showMenuOptions = showMenuOptions;
 
     // If the pointer is not warped to the center of the menu, we should not enter
     // turbo-mode right away.
-    this.input.deferredTurboMode = !options.warpMouse && options.centeredMode;
+    this.input.deferredTurboMode =
+      !showMenuOptions.warpMouse && showMenuOptions.centeredMode;
 
     this.input.update(this.getInitialMenuPosition());
     this.input.ignoreNextMotionEvents();
@@ -291,11 +339,11 @@ export class Menu extends EventEmitter {
 
     // If the menu is opened at the screen's center, we have to warp the mouse pointer to
     // the center of the menu.
-    if (this.options.warpMouse && this.options.centeredMode) {
+    if (this.showMenuOptions.warpMouse && this.showMenuOptions.centeredMode) {
       const position = this.getCenterItemPosition();
       const offset = {
-        x: Math.trunc(position.x - this.options.mousePosition.x),
-        y: Math.trunc(position.y - this.options.mousePosition.y),
+        x: Math.trunc(position.x - this.showMenuOptions.mousePosition.x),
+        y: Math.trunc(position.y - this.showMenuOptions.mousePosition.y),
       };
 
       this.emit('move-pointer', offset);
@@ -308,6 +356,9 @@ export class Menu extends EventEmitter {
   /** Hides the menu. */
   public hide() {
     this.container.classList.add('hidden');
+    this.hideTimeout = setTimeout(() => {
+      this.clear();
+    }, this.options.fadeOutDuration);
   }
 
   /** Removes all DOM elements from the menu and resets the root menu item. */
@@ -323,6 +374,26 @@ export class Menu extends EventEmitter {
     this.hoveredItem = null;
     this.draggedItem = null;
     this.selectionChain = [];
+  }
+
+  /**
+   * Sets the fade-in duration in milliseconds for future menu-open animations.
+   *
+   * @param duration The duration in milliseconds.
+   */
+  public setFadeInDuration(duration: number) {
+    this.options.fadeInDuration = duration;
+    this.container.style.setProperty('--fade-in-duration', `${duration}ms`);
+  }
+
+  /**
+   * Sets the fade-out duration in milliseconds for future menu-close animations.
+   *
+   * @param duration The duration in milliseconds.
+   */
+  public setFadeOutDuration(duration: number) {
+    this.options.fadeOutDuration = duration;
+    this.container.style.setProperty('--fade-out-duration', `${duration}ms`);
   }
 
   // --------------------------------------------------------------------- private methods
@@ -403,8 +474,8 @@ export class Menu extends EventEmitter {
       }
 
       if (item === this.root) {
-        const maxCenterTextSize = CENTER_RADIUS * 2.0;
-        const padding = CENTER_RADIUS * 0.1;
+        const maxCenterTextSize = this.options.centerDeadZone * 2.0;
+        const padding = this.options.centerDeadZone * 0.1;
         this.centerText = new CenterText(rootContainer, maxCenterTextSize - padding);
       }
     }
@@ -444,13 +515,13 @@ export class Menu extends EventEmitter {
     // center. There is the special case where we select the root item. In this case, we
     // simply position the root element at the mouse position.
     if (item === this.root) {
-      this.root.position = this.options.anchoredMode
+      this.root.position = this.showMenuOptions.anchoredMode
         ? this.getInitialMenuPosition()
         : this.input.absolutePosition;
     } else if (selectedParent) {
       const center = this.selectionChain[this.selectionChain.length - 1];
 
-      if (this.options.anchoredMode) {
+      if (this.showMenuOptions.anchoredMode) {
         this.root.position = math.add(this.root.position, center.position);
       } else {
         const offset = math.add(this.input.relativePosition, center.position);
@@ -458,16 +529,16 @@ export class Menu extends EventEmitter {
       }
     } else {
       // Compute the ideal position of the new item. The distance to the parent item is
-      // set to be at least PARENT_DISTANCE. This is to avoid that the menu is
-      // too close to the parent item. In anchored mode, the distance is set to
-      // PARENT_DISTANCE.
-      const distance = this.options.anchoredMode
-        ? PARENT_DISTANCE
-        : Math.max(PARENT_DISTANCE, this.input.distance);
+      // set to be at least this.options.minParentDistance. This is to avoid that the menu
+      // is too close to the parent item. In anchored mode, the distance is set to
+      // this.options.minParentDistance.
+      const distance = this.showMenuOptions.anchoredMode
+        ? this.options.minParentDistance
+        : Math.max(this.options.minParentDistance, this.input.distance);
 
       item.position = math.getDirection(item.angle, distance);
 
-      if (this.options.anchoredMode) {
+      if (this.showMenuOptions.anchoredMode) {
         this.root.position = math.subtract(this.root.position, item.position);
       } else {
         const offset = math.subtract(this.input.relativePosition, item.position);
@@ -493,7 +564,7 @@ export class Menu extends EventEmitter {
       const clampedPosition = math.clampToMonitor(
         position,
         this.theme.maxMenuRadius,
-        this.options.windowSize
+        this.showMenuOptions.windowSize
       );
 
       const offset = {
@@ -637,7 +708,7 @@ export class Menu extends EventEmitter {
     if (
       this.input.state === InputState.eDragging &&
       !this.draggedItem &&
-      this.input.distance > CENTER_RADIUS &&
+      this.input.distance > this.options.centerDeadZone &&
       this.hoveredItem
     ) {
       this.dragItem(this.hoveredItem);
@@ -648,7 +719,7 @@ export class Menu extends EventEmitter {
     if (
       this.input.state === InputState.eDragging &&
       this.draggedItem &&
-      this.input.distance < CENTER_RADIUS
+      this.input.distance < this.options.centerDeadZone
     ) {
       this.dragItem(null);
       this.updateConnectors();
@@ -687,7 +758,7 @@ export class Menu extends EventEmitter {
   private computeHoveredItem(): IRenderedMenuItem {
     // If the mouse is in the center of the menu, return the parent of the currently
     // selected item.
-    if (this.input.distance < CENTER_RADIUS) {
+    if (this.input.distance < this.options.centerDeadZone) {
       if (this.selectionChain.length > 1) {
         return this.selectionChain[this.selectionChain.length - 2];
       }
@@ -972,12 +1043,12 @@ export class Menu extends EventEmitter {
    */
   private getInitialMenuPosition() {
     return {
-      x: this.options.centeredMode
-        ? (this.options.windowSize.x / this.options.zoomFactor) * 0.5
-        : this.options.mousePosition.x,
-      y: this.options.centeredMode
-        ? (this.options.windowSize.y / this.options.zoomFactor) * 0.5
-        : this.options.mousePosition.y,
+      x: this.showMenuOptions.centeredMode
+        ? (this.showMenuOptions.windowSize.x / this.showMenuOptions.zoomFactor) * 0.5
+        : this.showMenuOptions.mousePosition.x,
+      y: this.showMenuOptions.centeredMode
+        ? (this.showMenuOptions.windowSize.y / this.showMenuOptions.zoomFactor) * 0.5
+        : this.showMenuOptions.mousePosition.y,
     };
   }
 }